requirements.txt
appdirs==1.4.4
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.37.4
idna==3.4
inflection==0.5.1
joblib==1.2.0
jsonlines==3.1.0
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.6.1
more-itertools==9.0.0
multitasking==0.0.11
numpy==1.23.3
openai==0.23.1
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
pandas-stubs==1.5.0.221003
patsy==0.5.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
Quandl==3.7.0
redditcleaner==1.1.2
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.0
six==1.16.0
sklearn==0.0
statsmodels==0.13.2
threadpoolctl==3.1.0
tqdm==4.64.1
types-pytz==2022.4.0.0
typing_extensions==4.4.0
urllib3==1.26.12
yfinance==0.1.77
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.7.8
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.3
cymem==2.0.6
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.0
fastjsonschema==2.16.2
filelock==3.8.0
huggingface-hub==0.10.0
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
murmurhash==1.0.8
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.6.1
nest-asyncio==1.5.6
nltk==3.7
notebook==6.4.12
numpy==1.23.3
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
preshed==3.0.7
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.9.2
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
regex==2022.9.13
requests==2.28.1
scikit-learn==1.1.2
Send2Trash==1.8.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
stack-data==0.5.1
terminado==0.16.0
textblob==0.17.1
thinc==8.1.3
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.13.1
torch==1.12.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
-e git+https://github.com/huggingface/transformers.git@c2b83d540e94c609b5ede3fbf6f374a577a747b7#egg=transformers
typer==0.4.2
typing_extensions==4.4.0
urllib3==1.26.12
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==4.0.3
zipp==3.8.1